Fix tests not working in travis but pass locally
